LoanApplication create loan method updated

+++ backend/loan-application/src/main/java/tech/zetapioneers/loan_application/controllers/LoanApplicationController.java
@@ -47,7 +47,7 @@ public class LoanApplicationController {
     @PostMapping
     public LoanApplicationResponse createLoan(@RequestBody LoanApplicationResponse loan) {
         System.out.println(loan.getAmount());
-        Long id = loanApplicationServiceImpl.addLoanApplication(loan);
+        Long id = (Long) loanApplicationServiceImpl.addLoanApplication(loan);
         loan.setId(id);
         loan.setRateOfInterest(loan.getPurpose().getInterestRate());
         return loan;
